+++ utils/login.py
@@ -56,7 +56,7 @@ def generarMenu(usuario):
         st.logo("assets/images/soccer-central.png", size="large")
 
         # Cargamos la tabla de usuarios
-        dfusuarios = pd.read_csv('usuarios.csv')
+        #dfusuarios = pd.read_csv('usuarios.csv')
 
         # Filtramos la tabla de usuarios
         #dfUsuario =dfusuarios[(dfusuarios['usuario']==usuario)]
@@ -70,9 +70,9 @@ def generarMenu(usuario):
         # Mostramos los enlaces de páginas
         st.page_link("app.py", label="Inicio", icon=":material/home:")
         st.subheader(":material/dashboard: Dashboard")
-        st.page_link("pages/player_report_from_layout.py", label="Reportes", icon=":material/picture_as_pdf:")
-        st.page_link("pages/player_assessments.py", label="Formularios de evaluacion", icon=":material/description:")
-        st.page_link("pages/sc_assessments.py", label="DEMO SC Evaluación", icon=":material/description:")
+        st.page_link("pages/player360.py", label="Player 360", icon=":material/contacts:")
+        st.page_link("pages/player_evaluation.py", label="Player Evaluation", icon=":material/description:")
+        st.page_link("pages/sc_player_evaluation.py", label="DEMO Player Evaluation", icon=":material/description:")
 
         st.subheader(":material/manage_accounts: Administrator")
 
@@ -101,7 +101,8 @@ def cerrarSesion():
     st.rerun()
 
 def generarLogin():
-    # 🔄 Restaurar sesión desde la URL si existe
+
+    #Restaurar sesión desde la URL si existe
     usuario_actual = st.query_params.get("user")
     if usuario_actual and "usuario" not in st.session_state:
         st.session_state['usuario'] = usuario_actual['username']